refactor(agent): dedupe sessionIdFromSse into a shared test helper
The in-band grida-session frame parser was copy-pasted byte-identical
across four suites (agent routes, agent events, e2e, live runtime) — a
frame-format change would have to land in four places. One test-only
helper at testing/sse.ts, anchored to GRIDA_SESSION_SSE_EVENT instead
of a hardcoded string. Found by the /simplify reuse pass.
