+++ packages/grida-ai-agent/src/e2e.test.ts
@@ -14,6 +14,7 @@
  *   - compaction  → the head is summarized + folded into the next user turn
  */
 import fs from "node:fs/promises";
+import { sessionIdFromSse } from "./testing/sse";
 import os from "node:os";
 import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
@@ -272,24 +273,6 @@ async function runTurn(opts: {
     expect(msgs.some((m) => m.role === "assistant")).toBe(true);
   });
   return sessionId;
-}
-
-/** Pull the session id out of a drained SSE body's `grida-session` frame. */
-function sessionIdFromSse(body: string): string {
-  for (const frame of body.split("\n\n")) {
-    if (!frame.startsWith("event: grida-session")) continue;
-    const dataLine = frame.split("\n").find((l) => l.startsWith("data:"));
-    if (!dataLine) continue;
-    try {
-      const parsed = JSON.parse(dataLine.slice("data:".length).trim()) as {
-        session_id?: string;
-      };
-      return parsed.session_id ?? "";
-    } catch {
-      return "";
-    }
-  }
-  return "";
 }
 
 async function waitFor(
